nice

+++ NumericalMethods/CentralDifference.cs
@@ -4,19 +4,19 @@ namespace NumericalMethods
 {
     class CentralDifference
     {
-        public static double Calculate(double x, double h, Func<double, double> f)
+        public static decimal Calculate(decimal x, decimal h, Func<decimal, decimal> f)
         {
             return (f(x + h) - f(x - h)) / (2 * h);
         }
 
-        public static double Extrapolate(double ei, double ei2)
+        public static decimal Extrapolate(decimal ei, decimal ei2)
         {
             return ei2 + ((ei2 - ei) / 3);
         }
 
-        public static double Extrapolate(double x, double h, Func<double, double> f)
+        public static decimal Extrapolate(decimal x, decimal h, Func<decimal, decimal> f)
         {
-            return Extrapolate(Calculate(x, h, f), Calculate(x, h / 2d, f));
+            return Extrapolate(Calculate(x, h, f), Calculate(x, h / 2m, f));
         }
     }
 }
